api/requirements.txt
django
djangorestframework
psycopg
djangorestframework-simplejwt